# docs/swagger.yaml
basePath: /
definitions:
  echoconfig.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  notification.NotificationDTO:
    properties:
      timer:
        $ref: '#/definitions/timermodel.Timer'
      type:
        $ref: '#/definitions/notification.NotificationType'
    type: object
  notification.NotificationType:
    enum:
    - notification_expired
    - notification_delete
    type: string
    x-enum-varnames:
    - Expired
    - Delete
  timerfields.Color:
    enum:
    - DEFAULT
    - RED
    - GREEN
    - BLUE
    - PURPLE
    - YELLOW
    type: string
    x-enum-varnames:
    - DEFAULT
    - RED
    - GREEN
    - BLUE
    - PURPLE
    - YELLOW
...
  timermodel.Timer:
    properties:
      color:
        $ref: '#/definitions/timerfields.Color'
      creator:
        type: integer
      description:
        type: string
      duration:
        type: integer
      endTime:
        type: string
      id:
        type: string
      isPaused:
        type: boolean
      name:
        type: string
      pauseTime:
        type: string
      type:
        type: string
      utc:
        type: integer
      withMusic:
        type: boolean
    type: object
...
host: localhost:12700
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Timer API Swagger
  version: "1.0"
paths:
  /notifications:
    delete:
      description: delete all user notifications
      parameters:
      - description: user id
        in: query
        name: vk_user_id
        required: true
        type: integer
      - description: you can add secret key to query for debug requests
        in: query
        name: debug
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
      summary: Delete
      tags:
      - notifications
    get:
      description: get user unread notifications, notifications include delete or
        expire timer
      parameters:
      - description: user id
        in: query
        name: vk_user_id
        required: true
        type: integer
      - description: you can add secret key to query for debug requests
        in: query
        name: debug
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/notification.NotificationDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
      summary: NotificationsByUser
      tags:
      - notifications
...
  /timers/:id/reset:
    patch:
      description: reset timer by timer id, only owner can reset timer, every subscriber
        (creator inclusive) will be send reset event, if timer is started, reset not
        pause, only update end time
      parameters:
      - description: user id
        in: query
        name: vk_user_id
        required: true
        type: integer
      - description: you can add secret key to query for debug requests
        in: query
        name: debug
        type: string
      - description: timer id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/timermodel.Timer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
      summary: ResetTimer
      tags:
      - timers
  /timers/:id/start:
    patch:
      description: start timer by timer id, only owner can start timer, every subscriber
        (creator inclusive) will be send start event
      parameters:
      - description: user id
        in: query
        name: vk_user_id
        required: true
        type: integer
      - description: you can add secret key to query for debug requests
        in: query
        name: debug
        type: string
      - description: timer id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/timermodel.Timer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
      summary: StartTimer
      tags:
      - timers
  /timers/:id/stop:
    patch:
      description: stop timer by timer id, only owner can stop timer, every subscriber
        (creator inclusive) will be send stop event
      parameters:
      - description: user id
        in: query
        name: vk_user_id
        required: true
        type: integer
      - description: you can add secret key to query for debug requests
        in: query
        name: debug
        type: string
      - description: timer id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
      summary: StopTimer
      tags:
      - timers
  /timers/:id/subscribe:
    post:
      description: subscribe user on timer by id, user will see timer in subscriptions,
        get events and notificaitons
      parameters:
      - description: user id
        in: query
        name: vk_user_id
        required: true
        type: integer
      - description: timer id
        in: path
        name: id
        required: true
        type: string
      - description: you can add secret key to query for debug requests
        in: query
        name: debug
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/timermodel.Timer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
      summary: Subscribe
      tags:
      - timers
...
  /timers/:id/unsubscribe:
    delete:
      description: unsubscribe user on timer by id, user wont see timer in subscriptions,
        get events and notificaitons
      parameters:
      - description: user id
        in: query
        name: vk_user_id
        required: true
        type: integer
      - description: timer id
        in: path
        name: id
        required: true
        type: string
      - description: you can add secret key to query for debug requests
        in: query
        name: debug
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echoconfig.ErrorResponse'
      summary: Unsubscribe
      tags:
      - timers
...
swagger: "2.0"